% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/designs.R, R/practical.R, R/profiles.R,
%   R/structural.R
\name{as_table.pkident_designs}
\alias{as_table.pkident_designs}
\alias{as_table.pkident_practical}
\alias{as_table.pkident_profiles}
\alias{as_table}
\alias{as_table.pkident_structural}
\title{Extract result tables}
\usage{
\method{as_table}{pkident_designs}(x, what = c("summary", "parameters"), ...)

\method{as_table}{pkident_practical}(x, what = c("parameters", "directions", "correlation"), ...)

\method{as_table}{pkident_profiles}(x, what = c("parameters", "profiles"), ...)

as_table(x, what, ...)

\method{as_table}{pkident_structural}(x, what = c("parameters", "combinations", "rank", "agreement"), ...)
}
\arguments{
\item{x}{A pkident result.}

\item{what}{Which table.}

\item{...}{Unused.}
}
\value{
A tibble.
}
\description{
Extract result tables
}
