% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/structural.R
\name{as_table}
\alias{as_table}
\alias{as_table.pkident_structural}
\title{Extract result tables}
\usage{
as_table(x, what, ...)

\method{as_table}{pkident_structural}(x, what = c("parameters", "combinations", "rank"), ...)
}
\arguments{
\item{x}{A pkident result.}

\item{what}{Which table.}

\item{...}{Unused.}
}
\value{
A tibble.
}
\description{
Extract result tables
}
